* Bug #42485: Do not delete joinscripts / schema files from Non-Docker Apps * Bug #42488: Mark database packages as user installed if already installed * Bug #42498: Call join scripts last when upgrading to Docker App
git-svn-id: svn+ssh://billy.knut.univention.de/var/univention/svn/dev/branches/ucs-4.2/ucs-4.2-0@72802 1b283449-9f2f-0410-b571-8eb228e1a901

+++ management/univention-appcenter/python/appcenter/actions/update.py
@@ -273,10 +273,11 @@ class Update(UniventionAppAction):
 				src = app.get_cache_file(file)
 				dest = update_files[file](app)
 				if not os.path.exists(src):
-					# remove files that do not exist on server anymore
-					if os.path.exists(dest):
-						self.log('Deleting obsolete app file %s' % dest)
-						os.unlink(dest)
+					if app.docker:
+						# remove files that do not exist on server anymore
+						if os.path.exists(dest):
+							self.log('Deleting obsolete app file %s' % dest)
+							os.unlink(dest)
 				else:
 					# update local files if downloaded
 					component_file = '%s.%s' % (app.component_id, file)
